aluno nao pode excluir documento de outro aluno.

+++ src/inscricoes/inscricoes.controller.ts
@@ -31,6 +31,8 @@ import { ProcessosSeletivosService } from 'src/processos-seletivos/processos-sel
 import { ProducaoCientificaService } from 'src/producao-cientifica/producao-cientifica.service';
 import { MailerService } from '@nestjs-modules/mailer';
 import { UsuariosService } from 'src/usuarios/usuarios.service';
+import { AlunosService } from 'src/alunos/alunos.service';
+
 
 @Controller('inscricoes')
 export class InscricoesController {
@@ -42,6 +44,7 @@ export class InscricoesController {
     private processosSeletivosService: ProcessosSeletivosService,
     private mailService: MailerService,
     private usuarioService: UsuariosService,
+    private alunosService: AlunosService,
   ) {}
 
   @Roles(Role.ALUNO)
@@ -172,9 +175,10 @@ export class InscricoesController {
 
   @Roles(Role.ALUNO)
   @UseGuards(JwtAuthGuard, RolesGuard)
-  @Delete('/producoes/:id')
+  @Delete(':inscricao_id/producoes/:id')
   async deleteProducao(
     @Param('id') id: string,
+    @Param('inscricao_id') inscricao_id: string,
     @Request() req,
   ) {
 
@@ -186,12 +190,30 @@ export class InscricoesController {
       );
     }
     else{
-      if (this.producaoCientificaService.deleteProducao({ id: Number(id) })){
-        return {
-          statusCode: HttpStatus.OK,
-          message: "Documento deletado"
+      const aluno = await this.alunosService.findAlunoByUserId(req.user.userId);
+      const inscricao = await this.inscricoesService.findOne(+inscricao_id);
+      if(!inscricao){
+        throw new HttpException(
+          'Inscricão não existe',
+          HttpStatus.NOT_FOUND,
+        );
+      }
+      
+      if(producao.inscricao_id != inscricao.id || inscricao.aluno_id != aluno.id) {
+        throw new HttpException(
+          'Ação não permitida.',
+          HttpStatus.FORBIDDEN,
+        );
+      }
+      else{
+        if (this.producaoCientificaService.deleteProducao({ id: Number(id) })){
+          return {
+            statusCode: HttpStatus.OK,
+            message: "Documento deletado"
+          }
         }
       }
+      
     }
     
   }
